Fix drag and drop in workspace thumbnails of overview
This was originally implemented by @mainland.

The ThumbnailsBox base class only uses the x coordinate to test whether a point
is in a workspace. This is not sufficient when we may have multiple rows of
thumbnails.

The new _withinWorkspace function also uses the y coordinate. This requires
overriding handleDragOver as well so that it passes both the x and y coordinates
to _withinWorkspace.

References https://github.com/mainland/gnome-shell-wsmatrix/commit/62be78e8
References https://github.com/mzur/gnome-shell-wsmatrix/issues/177